refactor(#3038): VoiceBargeInRuntime S3 — extract final-result playback cluster
S3 scope: extract the next cohesive VoiceBargeInRuntime subcluster from the S1 concern map without behavior changes. After S2, routing/dispatch remains the largest aggregate but is still an intertwined multi-subcluster routing surface. The remaining clean playback candidates are final-result playback and live-cut playback; final-result playback is the larger cohesive slice and moves cleanly: background completion summary speech, final answer TTS chunked playback, and spoken-result playback ownership bookkeeping.

Cluster boundary: moved `speak_voice_background_completion_summary`, `start_spoken_result_playback`, `clear_spoken_result_playback_if_current`, and `spawn_spoken_result_playback` verbatim into `src/services/discord/voice_barge_in/final_result_playback.rs` as an inherent impl child module. Live-cut playback helpers stay in the root for the next playback slice; progress playback and STT remain in their existing S2/S1 children.

No await point, lock scope, spawn, map-entry, cancellation, playback ownership, queue mutation, or channel-ordering behavior changed. The only moved-method visibility edits are `start_spoken_result_playback` and `clear_spoken_result_playback_if_current` as `pub(super)` so existing parent-module tests can keep exercising the helper state. No struct fields were widened. Existing cfg behavior is unchanged.

LoC accounting: `voice_barge_in.rs` production LoC drops 3932 -> 3710 (-222). New child `voice_barge_in/final_result_playback.rs` is 230 production LoC, below the 1000-line giant threshold, so only the root ratchet is lowered.

Token proof: normalized method-token comparison against HEAD is identical after visibility-only normalization; `start_spoken_result_playback` also normalizes the rustfmt trailing parameter comma introduced by the `pub(super)` line wrap. Per moved method:

| Method | Method tokens | Body tokens | Normalized SHA |
| --- | ---: | ---: | --- |
| `speak_voice_background_completion_summary` | 378 -> 378 | 329 -> 329 | `17eb90ee598c` |
| `start_spoken_result_playback` | 87 -> 87 | 69 -> 69 | `3c56366f5754` |
| `clear_spoken_result_playback_if_current` | 41 -> 41 | 25 -> 25 | `7089430ed82b` |
| `spawn_spoken_result_playback` | 851 -> 851 | 810 -> 810 | `4e8bf07faaeb` |
| **Total** | **1357 -> 1357** | **1233 -> 1233** | |

Remaining-map status from the S1 map after S3:

- routing/dispatch remains the largest aggregate and should be split by a narrower routing or dispatch ownership boundary, not moved wholesale.
- playback remaining in the root is now the live-cut subcluster: `reset_after_playback_start`, `reset_after_playback_start_with_owner`, `clear_playback`, `clear_playback_if_owner`, `observe_live_pcm_i16`, `monitor_for_channel`, `current_sensitivity`, and `update_existing_monitor_sensitivity`.
- TTS pipeline remains in the root: `set_runtime_tts_voice`, `synthesize_acknowledgement`, and `synthesize_progress_tts`.
- lifecycle/config and misc methods remain below the larger cohesive candidates.

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- cargo test --lib voice && cargo test --lib discord 2>&1 | tail -1
- set -o pipefail; cargo test --lib discord 2>&1 | tail -5
- python3 scripts/audit_maintainability.py --check
