+++ ARCHITECTURE.md
@@ -523,6 +523,7 @@ src/
 │   │   ├── turn_finalizer/
 │   │   │   └── cleanup.rs
 │   │   ├── voice_barge_in/
+│   │   │   ├── final_result_playback.rs
 │   │   │   ├── progress_playback.rs
 │   │   │   └── stt.rs
 │   │   ├── watchers/
