+++ scripts/audit_maintainability_giant_baseline.toml
@@ -97,7 +97,10 @@
 # #3038 VoiceBargeInRuntime S2: lowered 4350 -> 3932 (-418) as the progress
 # playback method cluster moved to `voice_barge_in/progress_playback.rs` (423
 # prod LoC, below the giant threshold). Locks in the shrink win.
-"src/services/discord/voice_barge_in.rs" = 3932
+# #3038 VoiceBargeInRuntime S3: lowered 3932 -> 3710 (-222) as the final-result
+# playback method cluster moved to `voice_barge_in/final_result_playback.rs`
+# (230 prod LoC, below the giant threshold). Locks in the shrink win.
+"src/services/discord/voice_barge_in.rs" = 3710
 # #3248: +60 for the gap-1 deploy-survivable-relay fix —
 # `reseed_watcher_owned_finalizer_ledger` + two guarded reattach call-sites that
 # re-register the watcher-owned turn in the post-restart finalizer ledger (a P1
